separated make_curve functions disabled /translateInclude option
