+++ curve_intersection/make_curve.ixx
@@ -1,0 +1,58 @@
+export module make_curve;
+
+import Point;
+import Bspline;
+
+export {
+	void make_curve01(Bspline& curve1, Bspline& curve2);
+
+	void make_curve02(Bspline& curve1, Bspline& curve2);
+
+	void make_curve03(Bspline& curve1, Bspline& curve2);
+
+	void make_curve04(Bspline& curve1, Bspline& curve2);
+
+	void make_curve05(Bspline& curve1, Bspline& curve2);
+
+	void make_curve06(Bspline& curve1, Bspline& curve2);
+
+	void make_curve07(Bspline& curve1, Bspline& curve2);
+
+	void make_curve08(Bspline& curve1, Bspline& curve2);
+
+	void make_curve09(Bspline& curve1, Bspline& curve2);
+
+	void make_curve10(Bspline& curve1, Bspline& curve2);
+
+	void make_curve11(Bspline& curve1, Bspline& curve2);
+
+	void make_curve12(Bspline& curve1, Bspline& curve2);
+
+	void make_curve13(Bspline& curve1, Bspline& curve2);
+
+	void make_curve14(Bspline& curve1, Bspline& curve2);
+
+	void make_curve15(Bspline& curve1, Bspline& curve2);
+
+	void make_curve16(Bspline& curve1, Bspline& curve2);
+
+	void make_curve17(Bspline& curve1, Bspline& curve2);
+
+	void make_curve18(Bspline& curve1, Bspline& curve2);
+
+	void make_curve19(Bspline& curve1, Bspline& curve2);
+
+	void make_curve20(Bspline& curve1, Bspline& curve2);
+
+	void make_curve21(Bspline& curve1, Bspline& curve2);
+
+	void make_curve22(Bspline& curve1, Bspline& curve2);
+
+	void make_curve23(Bspline& curve1, Bspline& curve2);
+
+	void make_curve24(Bspline& curve1, Bspline& curve2);
+
+	void make_curve25(Bspline& curve1, Bspline& curve2);
+
+	void make_curve26(Bspline& curve1, Bspline& curve2);
+}
